build(chakra): split chakra widgets and lib
