+++ packages/arco-lib/tsconfig.json
@@ -14,7 +14,6 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": false,
-    "emitDeclarationOnly": true,
     "jsx": "react-jsx",
     "outDir": "./lib",
     "baseUrl": "./",
