+++ packages/chakra-ui-lib/tsconfig.json
@@ -1,7 +1,6 @@
 {
   "extends": "../../config/tsconfig.base.json",
   "compilerOptions": {
-    "composite": true,
     "target": "ES6",
     "lib": ["DOM", "DOM.Iterable", "ESNext"],
     "outDir": "./lib",
