+++ packages/core/src/index.ts
@@ -8,3 +8,4 @@ export * from './version';
 export * from './slot';
 export * from './utilMethod';
 export * from './schema';
+export * from './metadata';
