+++ packages/runtime/index.html
@@ -14,7 +14,7 @@
       import { initSunmaoUI } from './src';
       import { ChakraProvider } from '@chakra-ui/react';
       import { sunmaoChakraUILib } from '@sunmao-ui/chakra-ui-lib';
-      import { ArcoDesignLib } from '@sunmao-ui/arco-lib';
+      import { ArcoDesignLib } from '@sunmao-ui/arco-lib/lib';
       import examples from '@example.json';
       import '@arco-design/web-react/dist/css/arco.css';
